Get rid of getFakeTokens
